+++ src/http/mod.rs
@@ -2,12 +2,11 @@
 
 use core::fmt;
 
-use alloc::string::String;
-
 #[cfg(feature = "macros")]
 pub mod macros;
 mod request;
 mod response;
+pub mod types;
 
 /// Enum for different supported HTTP versions
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
@@ -24,31 +23,6 @@ impl fmt::Display for HttpVersion {
         match self {
             HttpVersion::V1 => write!(f, "HTTP/1"),
             HttpVersion::V1_1 => write!(f, "HTTP/1.1"),
-        }
-    }
-}
-
-/// Content Type of the HTTP packet's body.
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub enum ContentType {
-    #[default]
-    /// text/plain
-    TextPlain,
-    /// application/json
-    ApplicationJson,
-    /// application/octet-stream
-    OctetStream,
-    /// Any other content type, as a string
-    Other(String),
-}
-
-impl fmt::Display for ContentType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ContentType::TextPlain => write!(f, "text/plain"),
-            ContentType::ApplicationJson => write!(f, "application/json"),
-            ContentType::OctetStream => write!(f, "application/octet-stream"),
-            ContentType::Other(s) => write!(f, "{s}"),
         }
     }
 }
